third_party/WebKit/LayoutTests/fast/dom/shadow/style-with-cat-expected.txt

Test for /shadow-deep/ combinator, http://crbug.com/309504.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS borderColorOf(getNodeInTreeOfTrees("host/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/host2/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("target2")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("target2")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/host2/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host//target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host//target")) is "rgb(0, 0, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host//host2/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/host2/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/host2/target")) is "rgb(0, 0, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/host2/host3/target")) is "rgb(0, 0, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/host2/target")) is "rgb(0, 0, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/host2/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/host2/host3/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/host2/target")) is "rgb(0, 0, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host/host2/host3/target")) is "rgb(0, 0, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host//host2/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("child")) is not "rgb(255, 0, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("inner/target")) is not "rgb(255, 0, 0)"
PASS borderColorOf(getNodeInTreeOfTrees("host//target")) is not "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
